+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - App Release Process
   - Updated roles endpoint to get roles
+- Connect Credential Management UI with new Issuer Component
 
 ## 2.0.0-RC3